Unicode BPE-tokenizer
This patch modified the BPE tokenizer implementation, so now it
is capable of joining unicode code points, and not only specific
byte pairs. This is necessary for a Gemma3 tokenizer that is
using unicode code points, unlike Tiktoken BPE.

+++ test/test_kernel_bmm.cc
@@ -64,10 +64,10 @@ TEST_CASE("Matmul single batch multiplication", "[kernel::bmm]")
 TEST_CASE("Matmul large 2d", "[!benchmark][kernel::bmm]")
 {
     metalchat::hardware_accelerator gpu0;
-    kernel::bmm<float> mm(gpu0);
+    kernel::bmm<bf16> mm(gpu0);
 
-    auto input1 = shared_tensor(full<float>({8, 2048}, 2.0));
-    auto input2 = shared_tensor(full<float>({2048, 128256}, 1.0));
+    auto input1 = shared_tensor(full<bf16>({8, 2048}, 2.0));
+    auto input2 = shared_tensor(full<bf16>({2048, 128256}, 1.0));
 
     BENCHMARK("multiply 128256 elements")
     {
